https://drive.google.com/drive/folders/1JVNwC0i4HEZD_mTot1w8SQuJGI9x0JlE
Food and Nutrition Quiz App is an educational mobile quiz app that covers various questions and answers about the science and interaction of nutrient and other substances in food in relation to maintenance, growth, reproduction and health of animals. users are to choose from the available options and summit. this displaces the scores and the right and wrong answers